Merge Orientation estimation branch

+++ Ganesha_II_V2/Core/Src/bmp581.c
@@ -1,4 +1,5 @@
 #include "bmp581.h"
+
 #include "bmp5.h"
 #include "bmp5_defs.h"
 #include "stm32h7xx_hal.h"
@@ -82,13 +83,13 @@ int8_t bmp581_update_data(struct BMP581 *bmp581, struct bmp5_sensor_data *data)
 static inline float calc_altitude_troposphere_msl(float pressure) {
     float exponent = (-GAS_CONSTANT * TROPOSPHERE_LAPSE_RATE) / (GRAVITY_ACCEL * AIR_MOLAR_MASS);
     float pressure_ratio = pressure / STANDARD_SEA_LEVEL_PRESSURE;
-    float power_term = pow(pressure_ratio, exponent) - 1;
+    float power_term = powf(pressure_ratio, exponent) - 1;
 
     return (STANDARD_SEA_LEVEL_TEMP / TROPOSPHERE_LAPSE_RATE) * power_term;
 }
 
 static inline float calc_altitude_lower_stratosphere_msl(float pressure) {
-    float log_ratio = log(pressure / TROPOPAUSE_PRESSURE);
+    float log_ratio = logf(pressure / TROPOPAUSE_PRESSURE);
     float scale_factor = (GAS_CONSTANT * STRATOSPHERE_BASE_TEMP) / (GRAVITY_ACCEL * AIR_MOLAR_MASS);
 
     return TROPOPAUSE_BASE_ALTITUDE - (scale_factor * log_ratio);
@@ -97,7 +98,7 @@ static inline float calc_altitude_lower_stratosphere_msl(float pressure) {
 static inline float calc_altitude_upper_stratosphere_msl(float pressure) {
     float exponent = (-GAS_CONSTANT * UPPER_STRATOSPHERE_LAPSE_RATE) / (GRAVITY_ACCEL * AIR_MOLAR_MASS);
     float pressure_ratio = pressure / STRATOSPHERE_MIDDLE_PRESSURE;
-    float power_term = pow(pressure_ratio, exponent) - 1;
+    float power_term = powf(pressure_ratio, exponent) - 1;
 
     return STRATOSPHERE_MIDDLE_BASE_ALTITUDE + (STRATOSPHERE_BASE_TEMP / UPPER_STRATOSPHERE_LAPSE_RATE) * power_term;
 }
